backend/requirements.txt
coverage==4.5.4
Django==3.0
django-filter==2.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
jsonschema==3.2.0
psycopg2-binary==2.8.4
requests==2.22.0
